requirements.txt
playwright>=1.42.0
yfinance>=0.2.37
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
